Fix issue where the first inserted user is not assigned the 'admin' role.
Modifies admin role assignment to check if the inserted user is the
first user based on sorted query using "createdAt".

Originally, the admin role assignment was done onCreateUser before
the user record was inserted into the db.  It assigned the 'admin' role
to a user if no other user existed in the db.  However, with the change
to insertUserDoc, the 'admin' assignment is done AFTER the first user
is inserted and the original query logic (check if any user exists) no
longer applies