Fix restricted domain registration regex
The current regex for checking restricted domains is not escaped and can
be easily tricked. For example, if the restricted domain is
"example.com", we currently accept "tester@hackedexample.com".

This change forces the full domain to be correct, and also escapes it to
prevent metacharacters from causing false matches.